Explosion implementation
Trying to do the explosions on the bombs when they hit.

I don't like this implementation - I started with a presenter and it
isn't working well.  In the middle I switched to a blackboard/event
based system and that's working better.

At this point when the bomb hits the explosion is off screen, the bomb
stays on screen, and then the game crashes.

Other than that it's working well :).