WIP: Start showing ads
This is a big change. For the first time I've got some cocos in the game
logic, in the form of a CocosDirectorAdapter.

You can create an AdDelegate with a director interface. In order to get
the interface to work with CCDirectorIOS we wrap it in an adapter. In
order to get that under test I've redefined CCDirector.m in the tests,
thereby allowing it to be instantiated via a link seam.