End the reign of OCDSpec2.
I started writing unit tests for navigation only to discover Apple broke
OCDSpec again. I'm done - I'm moving to Kiwi and I'll move the other
tests in on an as needed basis.

Maybe the solution will be to make my application logic an external
library that can be unit testing, and punting on the UI.